replace long series of if statements with loop and move to own util file
